Refactor ImageSaver
